Initial application build, containerisation and configuration.
Build for the application has been containerised with a `Dockerfile`, a docker compose build script has been created that provisions the images for the build of the golang application, the database PostgreSQL and the caching layer Redis. The latter two will be used as the primary datastore and "session" store respectively.

Ungraceful shutdown via SIGTERM & SIGINT signals have been handled appropriately.